Reduce memory load when installing big packages
Due to the UniversalFeedClient returning a non-seekable ConnectStream, ZipArchive internally creates a backing MemorySteam, thereby loading the entire package in
memory. This gives problems when handling big(1GB+) packages.

The solution suggested here is to have a simple command line option to save the package to disk, and then send the FileStream to UniversalPackage instead.
Since FileStream is seekable, ZipArchive will not create the backing MemoryStream.

The solution here is to save